requirements.txt
aiohttp
bottle
cheroot
PyJWT>=2.0.0
numpy
pymysql
sagemaker
six
SQLAlchemy==1.3.23
torch
transformers
werkzeug
boto3
yoyo-migrations
pandas==1.3.0
sacrebleu
sklearn
git+https://github.com/facebookresearch/dynalab.git@ab4bb29249eff7238ce3aae116707646e0d9429e
sentencepiece
spacy
textflint
